refactor(team): extract messaging endpoints into broker_messages.go
Slice 9. Moves the messaging HTTP surface, posters, visibility helpers,
and channel formatters (~798 lines) into a new broker_messages.go:
- handleMessages, handleNexNotifications, handlePostMessage, handleReactions
- PostSystemMessage, PostMessage, PostAutomationMessage, CreateRequest
- handleGetMessages
- Visibility helpers: messageInThread, normalizeMessageScope,
  messageMatchesViewerScope, messageVisibleToViewer,
  messageBelongsToViewerOutbox, messageBelongsToViewerInbox,
  messageRepliesToViewerThread, isOneOnOneDMMessage
- FormatChannelView, formatMessageUsageSuffix

Accessors that are interleaved with non-message methods (Messages,
ChannelMessages, AllMessages, PostInboundSurfaceMessage,
RecentHumanMessages) stay in broker.go for now.

Relocates 17+ messaging tests from broker_test.go into a new
broker_messages_test.go: format-channel-view, DM canonicalisation,
message kind+title, scope filtering, surface metadata, external queue,
post inbound, recent human messages, automation dedupe, focus mode
routing.

Adds three characterization tests:
- TestPostAutomationMessage_DedupesByEventID
- TestPostMessage_SetsTimestampAndChannel
- TestNormalizeMessageScope_KnownAndDefaults

Total package coverage rises from 63.2% to 63.3%.
